refactor(database): move dynamoSerializer to dbutil
