+++ internal/database/online/dynamodb/stream.go
@@ -3,6 +3,8 @@ package dynamodb
 import (
 	"context"
 
+	"github.com/oom-ai/oomstore/internal/database/dbutil"
+
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/feature/dynamodb/attributevalue"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
@@ -56,7 +58,7 @@ func (db *DB) Push(ctx context.Context, opt online.PushOpt) error {
 	item[opt.Entity.Name] = entityKeyValue
 
 	for i, feature := range opt.Features {
-		value, err := serializeByTag(opt.FeatureValues[i], feature.ValueType)
+		value, err := dbutil.SerializeByValueType(opt.FeatureValues[i], feature.ValueType, Backend)
 		if err != nil {
 			return err
 		}
